lint fixes

+++ src/WiFiServe.cpp
@@ -20,12 +20,12 @@ void startWiFiPortal() {
 
 void initWiFi() {
   // Connect to WiFi network
-  static uint8_t  MAC_STA[]                = {0,0,0,0,0,0};
+  static uint8_t MAC_STA[]                = {0,0,0,0,0,0};
   Serial.print("MAC[STA]");
   uint8_t *MAC  = WiFi.macAddress(MAC_STA);                   //get MAC address of STA interface
 
   wiFiMac = "";
-  for (uint8_t i = 0; i < sizeof(MAC)+2; ++i){
+  for (uint8_t i = 0; i < (uint8_t) (sizeof(MAC)+2); ++i){
     Serial.print(":");
     Serial.print(MAC[i],HEX);
     wiFiMac = wiFiMac + String(MAC[i], HEX);
